Remove DAO layer, bring everything into EsbMessageAdminServiceImpl. Fixes #118
