--- common/src/main/java/org/esbtools/message/admin/common/dao/MetadataDAOImpl.java
@@ -1,468 +1,0 @@
-/*
- Copyright 2015 esbtools Contributors and/or its affiliates.
-
- This file is part of esbtools.
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.esbtools.message.admin.common.dao;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-
-import javax.persistence.EntityManager;
-import javax.persistence.Query;
-
-import org.esbtools.message.admin.common.utility.ConversionUtility;
-import org.esbtools.message.admin.common.orm.MetadataEntity;
-import org.esbtools.message.admin.model.EsbMessage;
-import org.esbtools.message.admin.model.Header;
-import org.esbtools.message.admin.model.MetadataField;
-import org.esbtools.message.admin.model.MetadataResponse;
-import org.esbtools.message.admin.model.MetadataType;
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.slf4j.LoggerFactory;
-import org.slf4j.Logger;
-
-public class MetadataDAOImpl implements MetadataDAO {
-
-    private final EntityManager mgr;
-    private final AuditEventDAO auditDAO;
-    private static final Logger LOGGER=LoggerFactory.getLogger(MetadataDAOImpl.class);
-    private static transient Map<MetadataType, MetadataResponse> treeCache = new ConcurrentHashMap<>();
-    private static transient Map<String, List<String>> suggestionsCache = new ConcurrentHashMap<>();
-    private static final String ILLEGAL_ARGUMENT = "Illegal Argument:";
-    private static final String DEFAULT_USER = "someUser";
-    private static final String METADATA_KEY_TYPE = "metadata";
-    private static final String TYPE_PLACEHOLDER = "$TYPE";
-    private static final String METADATA_QUERY = "select f from MetadataEntity f where f.type = '" + TYPE_PLACEHOLDER + "'";
-
-    private Set<String> suggestedFields = new HashSet<>();
-    private List<String> resyncRestEndpoints = new ArrayList<>();
-
-    public MetadataDAOImpl(EntityManager mgr, AuditEventDAO auditDAO, JSONObject config) {
-        this.mgr=mgr;
-        this.auditDAO = auditDAO;
-        JSONArray suggestConfigs = (JSONArray) config.get("suggestedFields");
-        if(suggestConfigs!=null) {
-            for(int i=0;i<suggestConfigs.size();i++) {
-                suggestedFields.add(suggestConfigs.get(i).toString());
-            }
-        }
-        JSONArray resyncEndPoints = (JSONArray) config.get("resyncRestEndpoints");
-        if(resyncEndPoints!=null) {
-            for(Object endPoint: resyncEndPoints) {
-                resyncRestEndpoints.add(endPoint.toString());
-            }
-        }
-        if(resyncRestEndpoints.isEmpty()) {
-            throw new IllegalStateException("at least one resync rest end point needs to be configured");
-        }
-    }
-
-    private String getMetadataHash(MetadataType type) {
-        String hash = "";
-        if(type==MetadataType.SearchKeys || type==MetadataType.Entities) {
-            Query query = mgr.createQuery(METADATA_QUERY.replace(TYPE_PLACEHOLDER, type.toString()));
-
-            List<MetadataEntity> result = query.getResultList();
-            if(result!=null && !result.isEmpty()) {
-                hash = (String) result.get(0).getValue();
-            }
-        }
-        return hash;
-    }
-
-    private String markTreeDirty(MetadataType metadataType) {
-        String hash = null;
-
-        MetadataType type = metadataType.isSearchKeyType() ? MetadataType.SearchKeys : MetadataType.Entities;
-
-        Query query = mgr.createQuery(METADATA_QUERY.replace(TYPE_PLACEHOLDER, type.toString()));
-        List<MetadataEntity> result = query.getResultList();
-        if(result!=null && !result.isEmpty()) {
-            hash = UUID.randomUUID().toString();
-            result.get(0).setValue(hash);
-        }
-        return hash;
-    }
-
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.esbtools.message.admin.service.dao.MetadataDAO#getMetadataTree(org
-     * .esbtools.message.admin.model.MetadataType)
-     *
-     * fetch all metadata fields based on the the type of the tree requested.
-     * Then compute the tree from those fields and respond with the entire tree
-     * on the tree field, and the result field as null.
-     */
-    @Override
-    public MetadataResponse getMetadataTree(MetadataType type) {
-
-        MetadataResponse result;
-        if(type == MetadataType.Entities || type == MetadataType.SearchKeys) {
-            String hash = getMetadataHash(type);
-            if(treeCache.containsKey(type) && hash.contentEquals(treeCache.get(type).getHash())) {
-                return treeCache.get(type);
-            } else {
-                result = refreshCache(type, hash);
-            }
-        } else {
-            result = new MetadataResponse();
-            result.setErrorMessage(ILLEGAL_ARGUMENT + type + ", Expected: Entities or SearchKeys");
-        }
-        return result;
-    }
-
-    private MetadataResponse refreshCache(MetadataType type, String hash) {
-        MetadataResponse result;
-        result = new MetadataResponse();
-        String inClause = null;
-        if (type == MetadataType.Entities) {
-            inClause = "('Entities', 'Entity', 'System', 'SyncKey')";
-        } else {
-            inClause = "('SearchKeys', 'SearchKey', 'XPATH', 'Suggestion')";
-        }
-        if (inClause != null) {
-            Query query = mgr.createQuery("select f from MetadataEntity f where f.type in " + inClause);
-            List<MetadataEntity> queryResult = (List<MetadataEntity>) query.getResultList();
-            result.setTree(makeTree(queryResult));
-            result.setHash(hash);
-            treeCache.put(type, result);
-            if(type == MetadataType.SearchKeys) {
-                updateSuggestions(result.getTree());
-            }
-        }
-        return result;
-    }
-
-    /*
-     * given a list of metadata entities with parent ids, create a tree of
-     * Metadafields.
-     */
-    private static MetadataField makeTree(List<MetadataEntity> entities) {
-        MetadataField root = null;
-        Map<Long, MetadataField> map = new HashMap<>();
-        for (MetadataEntity entity : entities) {
-            MetadataField field = ConversionUtility.convertToMetadataField(entity);
-            if (entity.getType() == MetadataType.Entities || entity.getType() == MetadataType.SearchKeys) {
-                root = field;
-                root.setValue(entity.getType().toString());
-            }
-            map.put(field.getId(), field);
-        }
-        for (MetadataEntity entity : entities) {
-            MetadataField field = map.get(entity.getId());
-            MetadataField parent=null;
-            if (entity.getParentId().intValue() != -1) {
-                parent=map.get(entity.getParentId());
-                if(parent != null) {
-                    parent.addDescendant(field);
-                }
-            }
-        }
-        return root;
-    }
-
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.esbtools.message.admin.service.dao.MetadataDAO#addChildMetadataField
-     * (java.lang.Long, java.lang.String,
-     * org.esbtools.message.admin.service.model.MetadataType, java.lang.String)
-     *
-     * given a parent id, creates a metadata field and adds the new field as a
-     * child of the given parent. returns the entire tree of the metadata type
-     * and the parent field with all its children in the result field of the
-     * response.
-     */
-    @Override
-    public MetadataResponse addChildMetadataField(Long parentId, String name, MetadataType type, String value) {
-
-        MetadataResponse result = new MetadataResponse();
-        MetadataEntity curr = new MetadataEntity(type, name, value, parentId);
-        if (parentId == -1L) {
-            if (type != MetadataType.Entities && type != MetadataType.SearchKeys) {
-                result.setErrorMessage(ILLEGAL_ARGUMENT + type + ", If parent = -1, Expected: Entities or SearchKeys");
-            } else {
-                markTreeDirty(type);
-                mgr.persist(curr);
-                result = getMetadataTree(type);
-            }
-        } else {
-            MetadataField parent = getMetadataField(parentId);
-            if(parent==null) {
-                result.setErrorMessage(ILLEGAL_ARGUMENT + "parent "+parentId+ " not found!");
-            } else if (!curr.canBeChildOf(parent.getType())) {
-                result.setErrorMessage(ILLEGAL_ARGUMENT + type + " can not be a child of " + parent.getType());
-            } else {
-                markTreeDirty(type);
-                mgr.persist(curr);
-                result = createMetadataResult(parent);
-            }
-        }
-        auditDAO.save(DEFAULT_USER, "ADD", METADATA_KEY_TYPE, type.toString(), value, curr.toString());
-        return result;
-    }
-
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.esbtools.message.admin.service.dao.MetadataDAO#updateMetadataField
-     * (java.lang.Long, java.lang.String,
-     * org.esbtools.message.admin.service.MetadataType, java.lang.String)
-     *
-     * given a field id, overwrite the name, type and value of the metadata
-     * field return the entire metadata tree and the parent of the field being
-     * updated in the result field of the response.
-     */
-    @Override
-    public MetadataResponse updateMetadataField(Long id, String name, MetadataType type, String value) {
-
-        MetadataResponse result = new MetadataResponse();
-        MetadataEntity entity = mgr.find(MetadataEntity.class, id);
-
-        if (entity == null) {
-            result.setErrorMessage("Entity not found:" + id);
-        } else {
-            MetadataField parent = getMetadataField(entity.getParentId());
-            if (parent == null) {
-                result.setErrorMessage("Parent (" + entity.getParentId() + ") of Entity " + id + "not found!");
-            } else if (!entity.canBeChildOf(parent.getType())) {
-                result.setErrorMessage(type + " cannot be child of " + parent.getType());
-            } else {
-                entity.setName(name);
-                entity.setType(type);
-                entity.setValue(value);
-                markTreeDirty(type);
-                result = createMetadataResult(parent);
-            }
-        }
-        auditDAO.save(DEFAULT_USER, "UPDATE", METADATA_KEY_TYPE, type.toString(), value, entity.toString());
-        return result;
-
-    }
-
-    // keep children for history/ recovery, delete only current field
-    @Override
-    public MetadataResponse deleteMetadataField(Long id) {
-        MetadataResponse result = new MetadataResponse();
-        MetadataEntity entity = mgr.find(MetadataEntity.class, id);
-        mgr.remove(entity);
-        if (entity.getParentId() != -1L) {
-            MetadataField parent = getMetadataField(entity.getParentId());
-            markTreeDirty(parent.getType());
-            result = createMetadataResult(parent);
-        }
-        auditDAO.save(DEFAULT_USER, "DELETE", METADATA_KEY_TYPE, entity.getType().toString(), entity.getValue(), entity.toString());
-        return result;
-    }
-
-    /*
-     * given a Metadata field, create a MetadataResponse by looking up the
-     * entire tree. set the input field as the result in the MetadataResponse.
-     */
-    private MetadataResponse createMetadataResult(MetadataField field) {
-        MetadataResponse result = new MetadataResponse();
-        if (field.getType().isSyncKeyType()) {
-            result.setTree(getMetadataTree(MetadataType.Entities).getTree());
-        } else {
-            result.setTree(getMetadataTree(MetadataType.SearchKeys).getTree());
-        }
-        result.setResult(searchField(result.getTree(), field));
-        return result;
-    }
-
-    private MetadataField getMetadataField(Long id) {
-        MetadataEntity current = mgr.find(MetadataEntity.class, id);
-        return (current == null) ? null : ConversionUtility.convertToMetadataField(current);
-    }
-
-    /*
-     * DFS search
-     */
-    private MetadataField searchField(MetadataField tree, MetadataField field) {
-
-        MetadataField result = null;
-        if (tree != null && field != null) {
-            if (tree.getId().equals(field.getId())) {
-                return tree;
-            } else {
-                result = getMetadataField(tree, field);
-            }
-        }
-        return result;
-    }
-
-    private MetadataField getMetadataField(MetadataField tree, MetadataField field) {
-
-        for (MetadataField child : tree.getChildren()) {
-            MetadataField dfsResult = searchField(child, field);
-            if (dfsResult != null) {
-                return dfsResult;
-            }
-        }
-        return null;
-    }
-
-    @Override
-    public MetadataResponse sync(String entity, String system, String key, String... values) {
-
-        // create JMS Payload
-        StringBuilder message = new StringBuilder("<?xml version=\"1.0\" encoding=\"UTF-8\"?>");
-        message.append("<SyncRequest><EntityName>");
-        message.append(entity);
-        message.append("</EntityName><System>");
-        message.append(system);
-        message.append("</System><KeyName>");
-        message.append(key);
-        message.append("</KeyName>");
-        for(String value: values) {
-            if(value!=null && value.length()>0) {
-                message.append("<KeyValue>");
-                message.append(value);
-                message.append("</KeyValue>");
-            }
-        }
-        message.append("</SyncRequest>");
-        LOGGER.info("Initiating sync request: {}", message.toString());
-
-        auditDAO.save(DEFAULT_USER, "SYNC", METADATA_KEY_TYPE, entity, key, message.toString());
-
-        for(String restEndPoint: resyncRestEndpoints) {
-            try {
-                URL url = new URL(restEndPoint);
-                HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-                conn.setDoOutput(true);
-                conn.setRequestMethod("POST");
-                conn.setRequestProperty("Content-Type", "application/xml");
-                OutputStream os = conn.getOutputStream();
-                os.write(message.toString().getBytes("UTF-8"));
-                os.flush();
-                if (conn.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                    // status is Success by default
-                    return new MetadataResponse();
-                } else {
-                    // try another host
-                    LOGGER.warn("unable to send resync message, received Http response code:" +
-                            conn.getResponseCode() + " response message:" + conn.getResponseMessage() + " from:" + restEndPoint);
-                }
-                conn.disconnect();
-            } catch (MalformedURLException e) {
-                LOGGER.error(e.getMessage(), e);
-            } catch (IOException e) {
-                LOGGER.error(e.getMessage(), e);
-            }
-        }
-        MetadataResponse result = new MetadataResponse();
-        result.setErrorMessage("Unable to resync message");
-        return result;
-    }
-
-    @Override
-    public Map<String, List<String>> getSearchKeyValueSuggestions() {
-
-        // ensure cache exists and is upto date.
-        getMetadataTree(MetadataType.SearchKeys).getTree();
-        return suggestionsCache;
-    }
-
-    private void updateSuggestions(MetadataField searchKeysTree) {
-
-        Map<String, List<String>> newSuggestions = new HashMap<>();
-        if(searchKeysTree!=null && !searchKeysTree.getChildren().isEmpty()) {
-            for (MetadataField searchKey : searchKeysTree.getChildren()) {
-                addSuggestion(newSuggestions, searchKey);
-            }
-        }
-        suggestionsCache = newSuggestions;
-    }
-
-    private void addSuggestion(Map<String, List<String>> newSuggestions, MetadataField searchKey) {
-        if (suggestedFields.contains(searchKey.getValue())) {
-            List<String> values = new ArrayList<>();
-            for (MetadataField suggestion : searchKey.getSuggestions()) {
-                values.add(suggestion.getValue());
-            }
-            newSuggestions.put(searchKey.getValue(), values);
-        } else {
-            newSuggestions.put(searchKey.getValue(), null);
-        }
-    }
-
-    @Override
-    public void ensureSuggestionsArePresent(EsbMessage message, Map<String, List<String>> extractedHeaders) {
-
-        if(message.getHeaders()!=null) {
-            for(Header header:message.getHeaders()) {
-                if(suggestedFields.contains(header.getName())) {
-                    ensureSuggestionIsPresent(header.getName(), header.getValue());
-                }
-            }
-        }
-        for(String suggestedField: suggestedFields) {
-            List<String> extractedValues = extractedHeaders.get(suggestedField);
-            if(extractedValues!=null && !extractedValues.isEmpty()) {
-                for(String extractedValue: extractedValues) {
-                    ensureSuggestionIsPresent(suggestedField, extractedValue);
-                }
-            }
-        }
-    }
-
-    // should be called only for fields defined as suggested fields
-    private void ensureSuggestionIsPresent(String suggestedField, String suggestion) {
-        if(!suggestionsCache.containsKey(suggestedField)) {
-            Long parentId = treeCache.get(MetadataType.SearchKeys).getTree().getId();
-            addChildMetadataField(parentId, suggestedField, MetadataType.SearchKey, suggestedField);
-        }
-        Long searchKeyId = null;
-        if(!suggestionsCache.get(suggestedField).contains(suggestion)) {
-            searchKeyId = fetchSearchKeyId(suggestedField);
-            // fetch method can return null
-            if(searchKeyId!=null) {
-                addChildMetadataField(searchKeyId, suggestion, MetadataType.Suggestion, suggestion);
-            } else {
-                LOGGER.error("unable to add suggestion!");
-            }
-        }
-    }
-
-    private Long fetchSearchKeyId(String suggestedField) {
-        Query query = mgr.createQuery("select f from MetadataEntity f where f.value = :value");
-        query.setParameter("value", suggestedField);
-        List<MetadataEntity> queryResult = (List<MetadataEntity>) query.getResultList();
-        if (queryResult != null && !queryResult.isEmpty()) {
-            return queryResult.get(0).getId();
-        }
-        return null;
-    }
-
-}
